feat(blog): enforce post authorization using Gate in create, update, and delete actions

+++ app/Modules/Blog/Actions/DeletePost.php
@@ -12,7 +12,7 @@ class DeletePost {
     ) {
     }
     public function execute(Post $post) {
-        // Gate::authorize('delete' , $post);
+        Gate::authorize('delete', $post);
         return $this->service->delete($post);
     }
 }
